Fix compilation errors for examples and tests
- Added missing #include <cstdint> to string.hpp for uint32_t/uint64_t
- Fixed mapped_file template usage in example_mapped_file.cpp
- Undefined conflicting True/False macros in iso32000-1.hpp grammar

These fixes enable building with XTD_BUILD_EXAMPLES=TRUE and XTD_BUILD_TESTS=TRUE.

+++ examples/example_mapped_file.cpp
@@ -20,7 +20,7 @@ int main(){
     xtd::filesystem::path oFile = xtd::filesystem::temp_directory_path() /= "tmp_mapped_file.dat";
 
     {
-      xtd::mapped_file<-1> oMappedFile(oFile);
+      xtd::mapped_file<static_cast<size_t>(-1)> oMappedFile(oFile);
       auto oPerson = oMappedFile.get<person>(0);
       oPerson->age = 123;
       oPerson->ssn = 456;
@@ -28,7 +28,7 @@ int main(){
       strcpy(oPerson->last_name, "Marx");
     }
     {
-      xtd::mapped_file<-1> oMappedFile(oFile);
+      xtd::mapped_file<static_cast<size_t>(-1)> oMappedFile(oFile);
       auto oPerson = oMappedFile.get<person>(1);
       oPerson->age = 456;
       oPerson->ssn = 789;
@@ -36,7 +36,7 @@ int main(){
       strcpy(oPerson->last_name, "Marx");
     }
 
-    xtd::mapped_file<-1> oMappedFile(oFile);
+    xtd::mapped_file<static_cast<size_t>(-1)> oMappedFile(oFile);
     auto oGroucho = oMappedFile.get<person>(0);
     DBG(oGroucho->first_name, " ", oGroucho->last_name);
     auto oHarpo = oMappedFile.get<person>(1);
